Add files via upload
This commit features a clickable, resizable dartboard with animations and score calculation math. The inner and outer bullseye have been added and the math for the triple and double rings works. There is also a non-clickable outer ring that shows the single score value for that sector. The only thing left to change is to make the double and triple rings smaller relative to the rest of the dartboard.